backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.0
flask-cors==4.0.0
twilio==8.3.0